task

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -23,7 +23,7 @@ unsigned int binary_to_uint(const char *b)
 	while (*b != '\0')
 	{
 		if (*b == '0' || *b == '1')
-			r += (*b - 48) * a;
+			r += (unsigned int) (*b - 48) * a;
 		else
 			return (0);
 		b++;
